application/requirements.txt
celery
Django
django-celery-beat
django_celery_results
django-htmlmin
django-health-check
django-post_office
django-tinymce
djangorestframework
djangorestframework_simplejwt
filetype
gunicorn
psutil
python-dotenv
requests
sentry-sdk
Unidecode
urllib3
virtualenv
wheel